refactor: UI를 담당하는 로직 분리하기
View 폴더를 만들어 관리한다.
