+++ src/View/OutputView.js
@@ -4,6 +4,14 @@ const { Console } = require('@woowacourse/mission-utils');
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
  */
 const OutputView = {
+  printStartMessage() {
+    Console.print('다리 건너기 게임을 시작합니다.\n');
+  },
+
+  printErrorMessage(errorMsg) {
+    Console.print(errorMsg);
+  },
+
   /**
    * 현재까지 이동한 다리의 상태를 정해진 형식에 맞춰 출력한다.
    * <p>
@@ -12,6 +20,10 @@ const OutputView = {
   printMap([upBridge, downBridge]) {
     Console.print(`[ ${upBridge.join(' | ')} ]`);
     Console.print(`[ ${downBridge.join(' | ')} ]`);
+  },
+
+  printEndMessage(isFail) {
+    Console.print(`${!isFail ? '\n' : ''}최종 게임 결과`);
   },
 
   /**
